% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alluvial.R
\name{alluvial}
\alias{alluvial}
\title{Alluvial diagram}
\usage{
alluvial(..., freq, col = "gray", alpha = 0.5, border = 0, hide = FALSE,
  blocks = TRUE, layer, ranking = "zzout", ordering, gap.width = 0.05,
  xw = 0.1, cw = 0.1)
}
\arguments{
\item{...}{vectors or data frames to be combined into a single data frame 
(must have compatible numbers of rows).}

\item{freq}{numeric vector of frequencies, having the same length as the 
combined data frame.}

\item{col}{character vector of colors of ribbon colors.}

\item{alpha}{numeric vector of ribbon transparencies.}

\item{border}{character vector of ribbon border colors.}

\item{hide}{logical vector indicating which ribbons to hide.}

\item{blocks}{logical vector, having length compatible with the number of 
axes (factor variables), indicating which axes to represent as blocks 
(rather than stripes); or a character string indicating a predefined 
subset; options include "bookends".}

\item{layer}{numeric vector giving the order in which to draw the ribbons.}

\item{ranking}{character string indicating a predefined axis ranking
function; options include "rightward", "leftward", and "zzout" (default).}

\item{gap.width}{numeric, relative widths of inter-category gaps.}

\item{xw}{numeric, the distance from the set axis to the control points of 
each xspline.}

\item{cw}{numeric vector of axis widths.}
}
\value{
Nothing.
}
\description{
Drawing alluvial diagrams, also known as parallel set plots.
}
\details{
Alluvial diagrams have two principal uses: to track changing (relative) sizes
of data clusters over time, and to compare frequency data along several 
categorical variables (also use \code{mosaicplot} for two categorical 
variables). This package is still under development!
}
\examples{
# Titanic data
tit <- as.data.frame(Titanic)

# 2d
tit2d <- aggregate( Freq ~ Class + Survived, data=tit, sum)
alluvial( tit2d[,1:2], freq=tit2d$Freq, xw=0.0, alpha=0.8,
         gap.width=0.1, col= "steelblue", border="white",
         layer = tit2d$Survived != "Yes" )

alluvial( tit2d[,1:2], freq=tit2d$Freq, 
         hide=tit2d$Freq < 150,
         xw=0.0, alpha=0.8,
         gap.width=0.1, col= "steelblue", border="white",
         layer = tit2d$Survived != "Yes" )

# 3d
tit3d <- aggregate( Freq ~ Class + Sex + Survived, data=tit, sum)

alluvial(tit3d[,1:3], freq=tit3d$Freq, alpha=1, xw=0.2,
         col=ifelse( tit3d$Survived == "No", "red", "gray"),
         layer = tit3d$Sex != "Female",
         border="white")

# 4d
alluvial( tit[,1:4], freq=tit$Freq, border=NA,
         hide = tit$Freq < quantile(tit$Freq, .50),
         col=ifelse( tit$Class == "3rd" & tit$Sex == "Male", "red", "gray") )


# Possible blocks options
for (blocks in c(TRUE, FALSE, "bookends")) {
    
    # Elaborate alluvial diagram from main examples file
    alluvial( tit[, 1:4], freq = tit$Freq, border = NA,
              hide = tit$Freq < quantile(tit$Freq, .50),
              col = ifelse( tit$Class == "3rd" & tit$Sex == "Male",
                            "red", "gray" ),
              blocks = blocks )
}


### POLPAN data
if(FALSE)
{
# POLPAN panel sample
vnames <- grep("w[0-9]+", names(d), value=TRUE)
z <- aggregate( rep(1, nrow(d)), d[ , vnames ], sum)
z[,vnames] <- lapply( z[vnames], as.character)
# 1988 respondents
alluvial(z[,vnames], freq=z$x, col=ifelse(z$w1988, "red", "gray"))
# 2008 respondents
alluvial(z[,vnames], freq=z$x, col=ifelse(z$w2008, "red", "gray"))
}
}

