% Generated by roxygen2 (4.0.2): do not edit by hand
\name{alluvial_ts}
\alias{alluvial_ts}
\title{Alluvial diagram for multiple time series data}
\usage{
alluvial_ts(dat, wave = NA, ygap = 1, col = NA, plotdir = "up",
  rankup = T, xmargin = 1.1, lab.cex = 1, title = NA, title.cex = 1,
  xaxis.cex = 1, grid = F, grid.col = "grey80", grid.lwd = 1,
  leg.mode = T, leg.x = 0.1, leg.y = 0.9, leg.cex = 1,
  leg.col = "black", leg.lty = NA, leg.lwd = NA, leg.max = NA,
  lwd = NA, lwd.mult = 1, lwd.max = 1, ...)
}
\arguments{
\item{dat}{should be a 3 column data.frame with fields in order as per example below}

\item{wave}{wavyness of curves defined in terms of x axis data range - experiment to get this right}

\item{ygap}{gap between items on each y axis}

\item{col}{a single colour or vector of colours for categories when listed in alphabetical order}

\item{leg.mode}{if true legend plotted in largest data observation, otherwise custom coordinates (leg.x/y [0,1])

TODO: Add descriptions of remaining arguments above

TODO: Some details here}
}
\description{
This is a variant of alluvial diagram suitable for multiple
(cross-sectional) time series.
}

