% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowpoints.r
\name{flowpoints}
\alias{flowpoints}
\title{Compute endpoints of stripes (polygons or splines)}
\usage{
flowpoints(i, dat, freq, wid, ord)
}
\arguments{
\item{i}{numeric, an index between 1 and \code{ncol(dat)}.}

\item{dat}{data frame.}

\item{freq}{frequency vector.}

\item{wid}{numeric, gap between categories.}

\item{ord}{numeric vector giving the order in which to draw the stripes.}
}
\description{
Compute endpoints of stripes (polygons or splines)
}

