% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Polpan.R
\docType{data}
\name{Polpan}
\alias{Polpan}
\title{POLPAN voting flows}
\format{
Data frame with the following columns:
\describe{
\item{party2007}{Voting in 2007}
\item{party2011}{Voting in 2011}
\item{party2015}{Voting in 2015}
\item{n}{Counts}
}
}
\source{
\url{https://www.polpan.org}
}
\usage{
Polpan
}
\description{
Subset of POLPAN panel study.
}
\keyword{datasets}
