% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{Refugees}
\alias{Refugees}
\title{Refugees data}
\format{Describe columns
\describe{
\item{country}{country}
\item{year}{year}
\item{refugees}{number of refugees?}
}}
\source{
Where the data comes from?
}
\description{
What that data?
}
\references{
References, if any
}

