% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzout.R
\name{zzout}
\alias{zzout}
\title{Zigzag outward order}
\usage{
zzout(n, i)
}
\arguments{
\item{n}{numeric, a positive integer}

\item{i}{numeric, a positive integer at most \code{n}}
}
\description{
Order the numbers 1 through \code{n}, starting at some index and zigzagging outward.
}

