% Generated by roxygen2 (4.0.2): do not edit by hand
\name{alluvial}
\alias{alluvial}
\title{Alluvial diagram}
\usage{
alluvial(..., freq, col = "gray", border = 0, layer, hide = FALSE,
  alpha = 0.5, gap.width = 0.05, xw = 0.1, cw = 0.1)
}
\arguments{
\item{...}{vectors or data frames, all for the same number of observations}

\item{freq}{numeric, vector of frequencies of the same length as the number of observations}

\item{col}{vector of colors of the stripes}

\item{border}{vector of border colors for the stripes}

\item{layer}{numeric, order of drawing of the stripes}

\item{hide}{logical, should particular stripe be plotted}

\item{alpha}{numeric, vector of transparency of the stripes}

\item{gap.width}{numeric, relative width of inter-category gaps}

\item{xw}{numeric, the distance from the set axis to the control points of the xspline}

\item{cw}{numeric, width of the category axis}
}
\value{
Nothing
}
\description{
Drawing alluvial diagrams also known as parallel set plots.
}
\details{
Still under development!
}
\examples{
# Titanic data
tit <- as.data.frame(Titanic)

# 2d
tit2d <- aggregate( Freq ~ Class + Survived, data=tit, sum)
alluvial( tit2d[,1:2], freq=tit2d$Freq, xw=0.0, alpha=0.8,
         gap.width=0.1, col= "steelblue", border="white",
         layer = tit2d$Survived != "Yes" )

alluvial( tit2d[,1:2], freq=tit2d$Freq, 
         hide=tit2d$Freq < 150,
         xw=0.0, alpha=0.8,
         gap.width=0.1, col= "steelblue", border="white",
         layer = tit2d$Survived != "Yes" )

# 3d
tit3d <- aggregate( Freq ~ Class + Sex + Survived, data=tit, sum)

alluvial(tit3d[,1:3], freq=tit3d$Freq, alpha=1, xw=0.2,
         col=ifelse( tit3d$Survived == "No", "red", "gray"),
         layer = tit3d$Sex != "Female",
         border="white")

# 4d
alluvial( tit[,1:4], freq=tit$Freq, border=NA,
         hide = tit$Freq < quantile(tit$Freq, .50),
         col=ifelse( tit$Class == "3rd" & tit$Sex == "Male", "red", "gray") )


### POLPAN data
if(FALSE)
{
# POLPAN panel sample
vnames <- grep("w[0-9]+", names(d), value=TRUE)
z <- aggregate( rep(1, nrow(d)), d[ , vnames ], sum)
z[,vnames] <- lapply( z[vnames], as.character)
# 1988 respondents
alluvial(z[,vnames], freq=z$x, col=ifelse(z$w1988, "red", "gray"))
# 2008 respondents
alluvial(z[,vnames], freq=z$x, col=ifelse(z$w2008, "red", "gray"))
}
}

