% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Polpan.R
\docType{data}
\name{Polpan}
\alias{Polpan}
\title{POLPAN voting flows}
\format{
A data frame with 51 rows and 4 columns:
\describe{
\item{party2007}{Party voted for in the 2007 parliamentary election.}
\item{party2011}{Party voted for in the 2011 parliamentary election.}
\item{party2015}{Party voted for in the 2015 parliamentary election.}
\item{n}{Number of respondents with this combination of voting choices.}
}

Each of the \code{partyXXXX} columns takes one of the following values:
\code{"PO"} (Civic Platform / \emph{Platforma Obywatelska}),
\code{"PiS"} (Law and Justice / \emph{Prawo i Sprawiedliwosc}),
\code{"PSL"} (Polish People's Party / \emph{Polskie Stronnictwo Ludowe}),
\code{"Left"} (left-wing parties),
\code{"Other"}, or
\code{"not voted"}.
}
\source{
Polish Panel Study (POLPAN, \url{https://www.polpan.org}),
Marta Kołczyńska (\url{https://martakolczynska.com/})
and her \href{https://martakolczynska.com/post/polpan-voting-alluvial-plots/}{blog post}.
}
\usage{
data(Polpan)
}
\description{
Voting choices across three consecutive Polish parliamentary elections
(2007, 2011, 2015) reported by respondents of the Polish Panel Study
(POLPAN). Each row represents a unique combination of voting choices and
gives the number of respondents with that pattern.
}
\examples{
# color palette
p <- c("#E41A1C", "#377EB8", "#4DAF4A", "#984EA3", 
    "#FF7F00", "#FFFF33")

# How did the 2007 voters vote in subsequent elections?
alluvial(
    Polpan[,1:3],
    freq = Polpan$n,
    col = p[match(Polpan$party2007, sort(unique(Polpan$party2007)))]
)

# Remove flows with freqs less than 10
alluvial(
    Polpan[,1:3],
    freq = Polpan$n,
    col = p[match(Polpan$party2007, sort(unique(Polpan$party2007)))],
    hide = Polpan$n < 5
)


# How did the 2015 voters vote in preceding elections?
alluvial(
    Polpan[,1:3],
    freq = Polpan$n,
    col = p[match(Polpan$party2015, sort(unique(Polpan$party2007)))],
    hide = Polpan$n < 5
)



}
\keyword{datasets}
