% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Refugees.R
\docType{data}
\name{Refugees}
\alias{Refugees}
\title{Refugees data}
\format{
Data frame with the following columns:
\describe{
\item{country}{Country or territory of origin}
\item{year}{Year (2003-13)}
\item{refugees}{Persons recognized as refugees under the 1951 UN Convention, etc..}
}
}
\source{
\url{http://data.un.org/Data.aspx?d=UNHCR&f=indID\%3aType-Ref}
}
\description{
Top 10 countries/territories of origin (excluding "Various") for period
2003-13 of UNHCR statistics on "Persons recognized as refugees under the
1951 UN Convention/1967 Protocol, the 1969 OAU Convention, in accordance
with the UNHCR Statute, persons granted a complementary form of protection
and those granted temporary protection."
}
